# [Compiler+VM PoC] Allow compiler to generate instructions, retarget VM to use instructions