--- a/bbq/compiler/compiler_test.go
+++ b/bbq/compiler/compiler_test.go
@@ -43,7 +43,7 @@ func TestCompileRecursionFib(t *testing.T) {
   `)
 	require.NoError(t, err)
 
-	compiler := NewCompiler(checker.Program, checker.Elaboration)
+	compiler := NewBytecodeCompiler(checker.Program, checker.Elaboration)
 	program := compiler.Compile()
 
 	require.Len(t, program.Functions, 1)
@@ -114,7 +114,7 @@ func TestCompileImperativeFib(t *testing.T) {
   `)
 	require.NoError(t, err)
 
-	compiler := NewCompiler(checker.Program, checker.Elaboration)
+	compiler := NewBytecodeCompiler(checker.Program, checker.Elaboration)
 	program := compiler.Compile()
 
 	require.Len(t, program.Functions, 1)
@@ -203,7 +203,7 @@ func TestCompileBreak(t *testing.T) {
   `)
 	require.NoError(t, err)
 
-	compiler := NewCompiler(checker.Program, checker.Elaboration)
+	compiler := NewBytecodeCompiler(checker.Program, checker.Elaboration)
 	program := compiler.Compile()
 
 	require.Len(t, program.Functions, 1)
@@ -276,7 +276,7 @@ func TestCompileContinue(t *testing.T) {
   `)
 	require.NoError(t, err)
 
-	compiler := NewCompiler(checker.Program, checker.Elaboration)
+	compiler := NewBytecodeCompiler(checker.Program, checker.Elaboration)
 	program := compiler.Compile()
 
 	require.Len(t, program.Functions, 1)