--- a/bbq/vm/callframe.go
+++ b/bbq/vm/callframe.go
@@ -20,38 +20,12 @@ package vm
 
 import (
 	"github.com/onflow/cadence/bbq"
+	"github.com/onflow/cadence/bbq/opcode"
 )
 
 type callFrame struct {
 	executable   *ExecutableProgram
 	localsOffset uint16
 	localsCount  uint16
-	function     *bbq.Function
+	function     *bbq.Function[opcode.Instruction]
 }
-
-//
-//func (f *callFrame) getUint16() uint16 {
-//	first := f.function.Code[f.ip]
-//	last := f.function.Code[f.ip+1]
-//	f.ip += 2
-//	return uint16(first)<<8 | uint16(last)
-//}
-//
-//func (f *callFrame) getByte() byte {
-//	byt := f.function.Code[f.ip]
-//	f.ip++
-//	return byt
-//}
-//
-//func (f *callFrame) getBool() bool {
-//	byt := f.function.Code[f.ip]
-//	f.ip++
-//	return byt == 1
-//}
-//
-//func (f *callFrame) getString() string {
-//	strLen := f.getUint16()
-//	str := string(f.function.Code[f.ip : f.ip+strLen])
-//	f.ip += strLen
-//	return str
-//}