… bbq/vm/test/utils.go
… bbq/vm/test/vm_bench_test.go
… bbq/vm/test/vm_test.go