## Description

- Use one code generator instance in the compiler, and retarget it to the current function
- Provide constructors for compiling to bytecode and instructions
- Retarget the VM to instructions. This avoids decoding bytecode during execution
  
  This results in a bit of a performance improvement:

  Before:
  ```
  BenchmarkImperativeFib-10    	  558949	      2124 ns/op	      24 B/op	       3 allocs/op
  BenchmarkImperativeFib-10    	  557384	      2113 ns/op	      24 B/op	       3 allocs/op
  ```

  After:

  ```
  BenchmarkImperativeFib-10    	  626820	      1892 ns/op	      24 B/op	       3 allocs/op
  BenchmarkImperativeFib-10    	  631590	      1879 ns/op	      24 B/op	       3 allocs/op
  ```

______

<!-- Complete: -->

- [ ] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels